Add test executors

+++ Project/src/main/java/argssearch/retrieval/models/ConjunctiveRetrievalOnAllTables.java
@@ -98,9 +98,9 @@ public class ConjunctiveRetrievalOnAllTables {
             , final int tokenMinWeightDiscussion
             , final int tokenMinWeightPremise
             , final int tokenMinWeightArgument
-            , final int tokenWeightMultipierDiscussion
-            , final int tokenWeightMultipierPremise
-            , final int tokenWeightMultipierArgument
+            , final double tokenWeightMultipierDiscussion
+            , final double tokenWeightMultipierPremise
+            , final double tokenWeightMultipierArgument
             , final int limitDiscussion
             , final int limitPremise
             , final int limitArgument
@@ -111,15 +111,15 @@ public class ConjunctiveRetrievalOnAllTables {
         String tokenArray = preprocessedText.stream().map(this.cache::get).map(String::valueOf).collect(Collectors.joining(", ", "{", "}"));
 
         try{
-            this.query.setInt(1, tokenWeightMultipierArgument);
+            this.query.setDouble(1, tokenWeightMultipierArgument);
             this.query.setInt(2, tokenMinWeightArgument);
             this.query.setString(3, tokenArray);
             this.query.setInt(4, limitArgument);
-            this.query.setInt(5, tokenWeightMultipierPremise);
+            this.query.setDouble(5, tokenWeightMultipierPremise);
             this.query.setInt(6, tokenMinWeightPremise);
             this.query.setString(7, tokenArray);
             this.query.setInt(8, limitPremise);
-            this.query.setInt(9, tokenWeightMultipierDiscussion);
+            this.query.setDouble(9, tokenWeightMultipierDiscussion);
             this.query.setInt(10, tokenMinWeightDiscussion);
             this.query.setString(11, tokenArray);
             this.query.setInt(12, limitDiscussion);
